fixed products table and changed purchase type to double

+++ IMS/purchase.cs
@@ -112,7 +112,7 @@ namespace IMS
             da.Fill(dt);
             foreach (DataRow dr in dt.Rows)
             {
-                lblPurTotal.Text = Convert.ToString(Convert.ToInt32(txtQuantity.Text) * Convert.ToInt32(dr["Price"]));
+                lblPurTotal.Text = Convert.ToString(Convert.ToDouble(txtQuantity.Text) * Convert.ToDouble(dr["Price"]));
 
             }
